+++ www/front_src/src/Main/MainLoader.tsx
@@ -5,7 +5,7 @@ import { useTranslation } from 'react-i18next';
 import { Typography } from '@mui/material';
 import makeStyles from '@mui/styles/makeStyles';
 
-import logoCentreon from '../Navigation/Sidebar/Logo/centreon.png';
+import logoCentreon from '../assets/centreon.png';
 import { labelCentreonLogo } from '../Login/translatedLabels';
 
 import { labelCentreonIsLoading } from './translatedLabels';
