+++ www/front_src/src/Navigation/Sidebar/Logo/index.js
@@ -9,7 +9,7 @@ import React, { Component } from 'react';
 
 import clsx from 'clsx';
 
-import logo from './centreon.png';
+import logo from '../../../assets/centreon.png';
 import styles from './logo.scss';
 
 class Logo extends Component {
