+++ www/front_src/src/assets/centreon.png
